refactor(agent): split runtime cancel helpers
